Bug fixing, better example code, better test code

+++ Progetto/src/main/java/dao/jdbc/JDBCItemReviewDAO.java
@@ -29,17 +29,17 @@ public class JDBCItemReviewDAO extends JDBCDAO<ItemReview, Integer> implements I
     }
 
     @Override
-    public Long getCount() throws DAOException {
+    public Integer getCount() throws DAOException {
         try (PreparedStatement stm = CON.prepareStatement("SELECT COUNT(*) FROM ItemReview");) {
             ResultSet rs = stm.executeQuery();
             if (rs.next()) {
-                return rs.getLong(1);
+                return rs.getInt(1);
             }
         } catch (SQLException ex) {
             throw new DAOException("Impossible to count itemReviews", ex);
         }
 
-        return 0L;
+        return 0;
     }
 
     @Override
