+++ Progetto/src/main/java/dao/jdbc/JDBCUserDAO.java
@@ -28,17 +28,17 @@ public class JDBCUserDAO extends JDBCDAO<User, Integer> implements UserDAO{
     }
     
     @Override
-    public Long getCount() throws DAOException {
+    public Integer getCount() throws DAOException {
         try (PreparedStatement stm = CON.prepareStatement("SELECT COUNT(*) FROM User");) {
             ResultSet rs = stm.executeQuery();
             if (rs.next()) {
-                return rs.getLong(1);
+                return rs.getInt(1);
             }
         } catch (SQLException ex) {
             throw new DAOException("Impossible to count users", ex);
         }
 
-        return 0L;
+        return 0;
     }
 
     @Override
